fix: remove invalid properties from Vimeo component (#5176)
We accidentally added wrong properties by exporting a function with
iframe type.
